Tests/LibWeb/Text/expected/HTML/csp-frame-ancestors-srcdoc-inherited-header.txt

popup created: true
srcdoc script executed: false
iframe contentDocument is null: true
iframe body contains sentinel: false
